+++ packages/widgetbook/lib/src/core/args/single_arg.dart
@@ -1,12 +1,13 @@
 import '../fields/fields.dart';
 import '../framework/arg.dart';
 
-class SingleArg<T> extends Arg<T> with SingleFieldOnly {
+class SingleArg<T> extends StyledArg<T, SingleArgStyle> with SingleFieldOnly {
   SingleArg(
     super.value, {
     super.name,
     required this.values,
     this.labelBuilder,
+    super.style = const DropdownSingleArgStyle(),
   });
 
   final List<T> values;
@@ -14,21 +15,31 @@ class SingleArg<T> extends Arg<T> with SingleFieldOnly {
 
   @override
   Field<T> get field {
-    return ObjectDropdownField<T>(
-      name: 'value',
-      values: values,
-      initialValue: value,
-      labelBuilder: labelBuilder ?? ObjectDropdownField.defaultLabelBuilder,
-    );
+    return switch (style) {
+      DropdownSingleArgStyle() => ObjectDropdownField<T>(
+        name: 'value',
+        values: values,
+        initialValue: value,
+        labelBuilder: labelBuilder ?? ObjectDropdownField.defaultLabelBuilder,
+      ),
+      SegmentedSingleArgStyle() => ObjectSegmentedField<T>(
+        name: 'value',
+        values: values,
+        initialValue: value,
+        labelBuilder: labelBuilder ?? ObjectSegmentedField.defaultLabelBuilder,
+      ),
+    };
   }
 }
 
-class NullableSingleArg<T> extends Arg<T?> with SingleFieldOnly {
+class NullableSingleArg<T> extends StyledArg<T?, SingleArgStyle>
+    with SingleFieldOnly {
   NullableSingleArg(
     super.value, {
     super.name,
     required this.values,
     this.labelBuilder,
+    super.style = const DropdownSingleArgStyle(),
   });
 
   final List<T> values;
@@ -36,11 +47,31 @@ class NullableSingleArg<T> extends Arg<T?> with SingleFieldOnly {
 
   @override
   Field<T> get field {
-    return ObjectDropdownField<T>(
-      name: 'value',
-      values: values,
-      initialValue: value ?? values.first,
-      labelBuilder: labelBuilder ?? ObjectDropdownField.defaultLabelBuilder,
-    );
+    return switch (style) {
+      DropdownSingleArgStyle() => ObjectDropdownField<T>(
+        name: 'value',
+        values: values,
+        initialValue: value,
+        labelBuilder: labelBuilder ?? ObjectDropdownField.defaultLabelBuilder,
+      ),
+      SegmentedSingleArgStyle() => ObjectSegmentedField<T>(
+        name: 'value',
+        values: values,
+        initialValue: value,
+        labelBuilder: labelBuilder ?? ObjectSegmentedField.defaultLabelBuilder,
+      ),
+    };
   }
+}
+
+sealed class SingleArgStyle {
+  const SingleArgStyle();
+}
+
+final class DropdownSingleArgStyle extends SingleArgStyle {
+  const DropdownSingleArgStyle();
+}
+
+final class SegmentedSingleArgStyle extends SingleArgStyle {
+  const SegmentedSingleArgStyle();
 }